Refactor cV specific functionality into a separate library
