+++ src/DiagnosticSourceDemo/CorrelationVectorPropagation/HttpContextExtensions.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Http;
 
-namespace CorrelationPropagationDemos.DiagnosticSourceDemo
+namespace CorrelationVectorPropagation
 {
     public static class HttpContextExtensions
     {
