+++ src/DiagnosticSourceDemo/src/Controllers/ValuesController.cs
@@ -2,6 +2,7 @@
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Threading.Tasks;
+using CorrelationVectorPropagation;
 using Microsoft.AspNetCore.Mvc;
 
 namespace CorrelationPropagationDemos.DiagnosticSourceDemo
